Label encoding vs one hot encoding.
The way decision trees and random forest work using splitting logic, I was under the impression that label encoding would not be a problem, as we are anyway going to split the column. For eg: if we have gender as 'male', 'female' and 'other', with label encoding, it becomes 0,1,2 which is interpreted as 0<1<2. But since we we are going  to split the columns, I thought it didn't matter as it is the same thing we are going to split on 'male' or '0'. But when I tried both label and one hot encoding on the dataset, one hot encoding gave better accuracy and precision.